race timer uses integers now (broke compatibility with old file and sql databases)

+++ src/game/server/entities/character.h
@@ -146,8 +146,8 @@ private:
 	int m_LastSpeedup;
 	vec2 m_PrevPos;
 	
-	float CalculateFinishTime(float Time, vec2 PrevPos, vec2 Pos);
-	float CalculateStartAddTime(vec2 PrevPos, vec2 Pos);
+	int CalculateFinishTime(int Time, vec2 PrevPos, vec2 Pos);
+	int CalculateStartAddTime(vec2 PrevPos, vec2 Pos);
 };
 
 #endif
